Update ensemble_of_tracers.jl
This now all runs fine. It is not super fast but it is doing a lot (setting up 10 QG probs, 10 AD probs stepping them all forward and plotting etc) so I am not worried about it being a little on the slow side. There may be a better way so that only one `QG_prob` is needed but for now this is what I came up with!